refactor & restructure with graph document extraction
