+++ pyproject.toml
@@ -7,9 +7,14 @@ authors = [
 ]
 license = { file = "LICENSE" }
 readme = "README.md"
-requires-python = ">=3.13"
+requires-python = ">=3.13,<4.0"
 dependencies = [
-    "streamlit (>=1.46.1,<2.0.0)"
+    "streamlit (>=1.46.1,<2.0.0)",
+    "langchain (>=0.3.26,<0.4.0)",
+    "langchain-community (>=0.3.27,<0.4.0)",
+    "pypdf (>=5.8.0,<6.0.0)",
+    "langchain-experimental (>=0.3.4,<0.4.0)",
+    "langchain-ollama (>=0.3.5,<0.4.0)"
 ]
 
 
@@ -22,4 +27,5 @@ pre-commit = "^4.2.0"
 black = "^25.1.0"
 ruff = "^0.12.3"
 pytest = "^8.4.1"
+ipython = "^9.4.0"
